requirements.txt
Pillow==9.5.0
PyQt5==5.15.9
requests==2.31.0